Modifiquei os parametros para Integer, mas é longo.

+++ src/main/java/br/edu/iftm/vitrino/repository/ProdutoRepository.java
@@ -12,8 +12,8 @@ public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 
     Optional<Produto> findByNome(String nome);
 
-    Optional<Produto> findById(Long id);
+    Optional<Produto> findById(Integer id);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
 }
